requirements.txt
django==1.11
django-bootstrap3==11.1
Pillow
opencv-python
dlib==19.7.0
face_recognition==1.2.3 
africastalking==1.1.9